llm/requirements.txt
torch
transformers
python-dotenv
